backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pymongo==4.5.0
motor==3.3.1
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.7
Pillow==10.0.1
numpy==1.26.0
scikit-learn==1.3.1
matplotlib==3.8.0